add more ddnet9 sources

+++ arm9/src/engine/shared/snapshot.cpp
@@ -427,7 +427,7 @@ void CSnapshotStorage::PurgeUntil(int Tick)
 	m_pLast = 0;
 }
 
-void CSnapshotStorage::Add(int Tick, unsigned Tagtime, int DataSize, void *pData, int CreateAlt)
+void CSnapshotStorage::Add(int Tick, int64 Tagtime, int DataSize, void *pData, int CreateAlt)
 {
 	// allocate memory for holder + snapshot_data
 	int TotalSize = sizeof(CHolder)+DataSize;
@@ -463,7 +463,7 @@ void CSnapshotStorage::Add(int Tick, unsigned Tagtime, int DataSize, void *pData
 	m_pLast = pHolder;
 }
 
-int CSnapshotStorage::Get(int Tick, unsigned *pTagtime, CSnapshot **ppData, CSnapshot **ppAltData)
+int CSnapshotStorage::Get(int Tick, int64 *pTagtime, CSnapshot **ppData, CSnapshot **ppAltData)
 {
 	CHolder *pHolder = m_pFirst;
